scheduling/pyproject.toml
[tool.poetry]
name = "scheduling"
version = "0.1.0"
description = ""
authors = ["Simon <berndts.tin21@student.dhbw-heidenheim.de>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
schedule = "^1.2.0"
redis = "^4.6.0"
cassandra-driver = "^3.28.0"
docker-py = "^1.10.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
